requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
python-multipart==0.0.6
torch==2.9.1
torchvision==0.24.1
pillow==12.0.0
numpy==2.3.5
pandas==2.3.3
scikit-learn==1.7.2
matplotlib==3.10.7
tensorboard==2.20.0
tqdm==4.67.1
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
